refactor: Complete architecture reorganization with expanded domain APIs
- Restructure codebase into clear api/domain/infrastructure layers
- Add domain-specific APIs: Stock, Corp, Funds, Macro, Price
- Expand documentation with architecture, migration, and usage guides
- Add comprehensive examples for API usage
- Separate infrastructure clients for Yahoo and FRED
- Add utility helpers for caching and technical indicators
- Clean up old plan and integration test files

+++ src/main/kotlin/com/ulalax/ufc/domain/chart/ChartDataResponse.kt
@@ -216,8 +216,8 @@ data class DividendEvent(
 @Serializable
 data class SplitEvent(
     val date: Long? = null,
-    val numerator: Int? = null,
-    val denominator: Int? = null,
+    val numerator: Double? = null,
+    val denominator: Double? = null,
     val splitRatio: String? = null
 )
